requirements.txt
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0
TA-Lib>=0.4.25
ccxt>=3.0.0
websocket-client>=1.4.0
requests>=2.28.0
python-binance>=1.0.15
matplotlib>=3.6.0
seaborn>=0.11.0
tensorflow>=2.10.0
yfinance>=0.2.0
python-dotenv>=1.0.0
pydantic>=1.10.0
fastapi>=0.95.0
uvicorn>=0.20.0
schedule>=1.2.0
redis>=4.5.0